requirements.txt
aioredis==1.3.1
asgiref==3.2.3
asn1crypto==1.2.0
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
certifi==2019.11.28
cffi==1.13.2
channels==2.3.1
channels-redis==2.4.1
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
daphne==2.4.0
Django==2.2.8
django-cors-headers==3.2.0
django-filter==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
openapi-codec==1.3.2
pkg-resources==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.4.4
pyOpenSSL==19.1.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.2
requests==2.22.0
service-identity==18.1.0
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==18.8.1
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.7
wrapt==1.11.2
zope.interface==4.7.1
